working allocator

+++ src/db/file.h
@@ -12,16 +12,16 @@ int open_file(const char *filename);
 
 int close_file(int file_descriptor);
 
-int change_file_size(int fd, long new_size);
+int change_file_size(int fd, uint64_t new_size);
 
 int sync_file(int fd);
 
-int munmap_file(void *file_data_pointer, size_t file_size);
+int munmap_file(void *file_data_pointer, uint64_t file_size);
 
-int mmap_file(int fd, void **file_data_pointer, off_t offset, size_t size);
+int mmap_file(int fd, void **file_data_pointer, uint64_t offset, uint64_t size);
 
 int delete_file(const char *filename);
 
-size_t get_file_size(int fd);
+uint64_t get_file_size(int fd);
 
 #endif //LLP_DATABASE_FILE_H
